+++ src/ECMAScript.TDesign/TLayout.cs
@@ -5,9 +5,6 @@ namespace ECMAScript.TDesign;
 [VueLibraryComponent("tdesign-vue-next", "Layout")]
 [VueLibraryStyle("tdesign-vue-next/es/style/index.css")]
 [VueLibraryPluginRequirement("tdesign")]
-[VueProp(nameof(CssClass), Name = "class")]
-[VueProp(nameof(CssStyle), Name = "style")]
-[VueSlot(nameof(ChildContent), IsDefault = true)]
 public sealed class TLayout : TDesignContentComponentBase
 {
     [Parameter]
@@ -17,9 +14,6 @@ public sealed class TLayout : TDesignContentComponentBase
 [VueLibraryComponent("tdesign-vue-next", "Header")]
 [VueLibraryStyle("tdesign-vue-next/es/style/index.css")]
 [VueLibraryPluginRequirement("tdesign")]
-[VueProp(nameof(CssClass), Name = "class")]
-[VueProp(nameof(CssStyle), Name = "style")]
-[VueSlot(nameof(ChildContent), IsDefault = true)]
 public sealed class THeader : TDesignContentComponentBase
 {
     [Parameter]
@@ -29,9 +23,6 @@ public sealed class THeader : TDesignContentComponentBase
 [VueLibraryComponent("tdesign-vue-next", "Aside")]
 [VueLibraryStyle("tdesign-vue-next/es/style/index.css")]
 [VueLibraryPluginRequirement("tdesign")]
-[VueProp(nameof(CssClass), Name = "class")]
-[VueProp(nameof(CssStyle), Name = "style")]
-[VueSlot(nameof(ChildContent), IsDefault = true)]
 public sealed class TAside : TDesignContentComponentBase
 {
     [Parameter]
@@ -41,22 +32,16 @@ public sealed class TAside : TDesignContentComponentBase
 [VueLibraryComponent("tdesign-vue-next", "Content")]
 [VueLibraryStyle("tdesign-vue-next/es/style/index.css")]
 [VueLibraryPluginRequirement("tdesign")]
-[VueProp(nameof(CssClass), Name = "class")]
-[VueProp(nameof(CssStyle), Name = "style")]
-[VueProp(nameof(Text), Name = "content")]
-[VueSlot(nameof(ChildContent), IsDefault = true)]
 public sealed class TContent : TDesignContentComponentBase
 {
     [Parameter]
+    [ECMAScriptName("content")]
     public string? Text { get; set; }
 }
 
 [VueLibraryComponent("tdesign-vue-next", "Footer")]
 [VueLibraryStyle("tdesign-vue-next/es/style/index.css")]
 [VueLibraryPluginRequirement("tdesign")]
-[VueProp(nameof(CssClass), Name = "class")]
-[VueProp(nameof(CssStyle), Name = "style")]
-[VueSlot(nameof(ChildContent), IsDefault = true)]
 public sealed class TFooter : TDesignContentComponentBase
 {
     [Parameter]